Hoofdstuk 3 afgemaakt

+++ src/Matrix.h
@@ -7,16 +7,16 @@ class Matrix
 {
 protected:
     unsigned int MaxLength = 1;
-    std::vector<double> data;
+    std::vector<float> data;
 public:
     Matrix(unsigned int size);
 
     virtual void Print();
     virtual void PrintFormatted();
     virtual int GetLength() const;
-    virtual void Fill(std::vector<double>& values);
-    virtual double operator() (unsigned int row, unsigned int column) const;
-    virtual double& operator() (unsigned int row, unsigned int column);
+    virtual void Fill(std::vector<float>& values);
+    virtual float operator() (unsigned int row, unsigned int column) const;
+    virtual float& operator() (unsigned int row, unsigned int column);
     virtual bool operator== (const Matrix& other) const;
     virtual bool operator!= (const Matrix& other) const;
     ~Matrix() = default;
