+++ src/Matrix2.cpp
@@ -5,7 +5,7 @@ Matrix2::Matrix2() : Matrix(2)
 
 }
 
-double Matrix2::Determinant()
+float Matrix2::Determinant()
 {
     return (*this)(0, 0) * (*this)(1, 1) - (*this)(0, 1) * (*this)(1, 0);
 }
