+++ src/Matrix3.h
@@ -1,7 +1,6 @@
 #ifndef MATRIX3_H
 #define MATRIX3_H
 
-#include "Matrix.h"
 #include "Matrix2.h"
 
 class Matrix3 : public virtual Matrix
@@ -9,9 +8,9 @@ class Matrix3 : public virtual Matrix
 public:
     Matrix3();
     Matrix2 Submatrix(const unsigned int row, const unsigned int column) const;
-    double Minor(const unsigned int row, const unsigned int column) const;
-    double Cofactor(const unsigned int row, const unsigned int column) const;
-    double Determinant() const;
+    float Minor(const unsigned int row, const unsigned int column) const;
+    float Cofactor(const unsigned int row, const unsigned int column) const;
+    float Determinant() const;
 };
